pyproject.toml
[project]
name = "app-request"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "pillow>=11.2.1",
    "requests>=2.32.3",
]
